# backend/swagger.yaml
openapi: 3.0.0
info:
    description: |
        API du site Kasa. Ici on retrouvera les différentes routes pour la récupération/modification/suppression d'un hébérgement. Mais aussi pour l'authentification.
    version: "1.0.0"
    title: Kasa App
    contact:
        email: quentindumon352a@gmail.com
    license:
        name: Apache 2.0
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
    # Added by API Auto Mocking Plugin
    # Added by API Auto Mocking Plugin
    - description: SwaggerHub API Auto Mocking
      url: https://virtserver.swaggerhub.com/QUENTINDUMON352A/KasaApp/1.0.0
    - url: "http://localhost:3001"
...
paths:
...
    /api/housing/{id}:
        parameters:
            - name: id
              in: path
              description: ID de l'hébergement à récupérer
              required: true
              schema:
                type: string
        get:
          tags:
            - housing
          summary: Récupèration d'un hébergement avec son id spécifié dans les paramètres de la requête
          operationId: findOneHousing
          responses:
            "200":
              description: Opération réussite ✅
              content:
                application/json:
                  schema:
                    type: object
                    properties:
                      title:
                        type: string
                      cover:
                        type: string
                      pictures:
                        type: array
                        items:
                          type: string
                      description:
                        type: string
                      host:
                        type: object
                        properties:
                          name:
                            type: string
                          picture:
                            type: string
                          rating:
                            type: number
                      equipements:
                        type: array
                        items:
                          type: string
                      tags:
                        type: array
                        items:
                          type: string
        
            "400":
              description: Requête invalide ❌
            "404":
              description: Hébergement non trouvé 🤷‍♂
        delete:
            tags:
                - housing
            summary: Suppression d'un hébérgement 🛑
            description: Permet de supprimer un hébérgement avec l'id passé en paramètres ⚙️⚙️⚙️ de la requête ❌
            operationId: deleteHousing
            parameters:
                - name: token
                  in: header
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: Réquête réussie ✅
                "400":
                    description: Requête invalide ⛔️
                "401":
                    description: Droits refusés 🛑️
                "404":
                    description: Hébérgement non trouvé 🤷‍♂
            security:
                - bearerAuth: []
        put:
            tags:
                - housing
            summary: Mise à jour d'un hébérgement existant ♻️
            description: Permet de mettre à jour l'hébérgement spécifié dans les paramètres ⚙️⚙️⚙️ de la requète 👌
            operationId: updateHousing
            parameters:
                - name: token
                  in: header
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: Réquête réussie ✅
                "400":
                    description: Requête invalide ⛔️
                "401":
                    description: Droits refusés 🛑️
                "404":
                    description: Hébérgement non trouvé 🤷‍♂
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                book:
                                    type: string
                                image:
                                    type: string
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer